@@ -16,7 +16,7 @@
     <version>0.0.1-SNAPSHOT</version>
     <name>lyh-OJ</name>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
     <dependencies>
         <dependency>
@@ -132,6 +145,8 @@
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <configuration>
+                    <source>17</source>  <!-- 与 JDK 版本一致 -->
+                    <target>17</target>
                     <excludes>
                         <exclude>
                             <groupId>org.projectlombok</groupId>
