@@ -1,13 +1,14 @@
 package com.guanghan.lyhoj.judge.codesandbox.impl;
 
+import com.guanghan.lyhoj.judge.codesandbox.CodeSandboxTemplate;
 import com.guanghan.lyhoj.judge.codesandbox.CodesandBox;
 import com.guanghan.lyhoj.judge.codesandbox.model.ExecuteCodeRequest;
 import com.guanghan.lyhoj.judge.codesandbox.model.ExecuteCodeResponse;
 
-public class thirdCodeSandBox implements CodesandBox {
+public class CppCodeSandBox extends CodeSandboxTemplate  {
     @Override
-    public ExecuteCodeResponse ExecuteCode(ExecuteCodeRequest executeCodeRequest) {
-        System.out.println("第三方");
+    public ExecuteCodeResponse executeCode(ExecuteCodeRequest executeCodeRequest) {
+        System.out.println("cpp");
         return null;
     }
 }
